+++ src/test/resources/org/kohsuke/github/ReleaseTest/wiremock/testCreateSimpleRelease/__files/repos_hub4j-test-org_testcreaterelease-2.json
@@ -66,7 +66,7 @@
   "deployments_url": "https://api.github.com/repos/hub4j-test-org/testCreateRelease/deployments",
   "created_at": "2021-06-10T01:25:59Z",
   "updated_at": "2021-06-10T01:31:14Z",
-  "pushed_at": "2021-06-10T15:39:13Z",
+  "pushed_at": "2021-06-10T15:50:44Z",
   "git_url": "git://github.com/hub4j-test-org/testCreateRelease.git",
   "ssh_url": "git@github.com:hub4j-test-org/testCreateRelease.git",
   "clone_url": "https://github.com/hub4j-test-org/testCreateRelease.git",
