+++ src/test/resources/org/kohsuke/github/ReleaseTest/wiremock/testCreateSimpleRelease/__files/user-1.json
