+++ src/test/resources/org/kohsuke/github/ReleaseTest/wiremock/testCreateSimpleRelease/mappings/repos_hub4j-test-org_testcreaterelease-2.json
@@ -1,8 +1,8 @@
 {
-  "id": "c1b5c491-edb4-4985-90b0-3b027fd3a6a7",
-  "name": "user",
+  "id": "7cb5aa21-2b0d-4139-be8c-c12816556e50",
+  "name": "repos_hub4j-test-org_testcreaterelease",
   "request": {
-    "url": "/user",
+    "url": "/repos/hub4j-test-org/testCreateRelease",
     "method": "GET",
     "headers": {
       "Accept": {
@@ -12,25 +12,25 @@
   },
   "response": {
     "status": 200,
-    "bodyFileName": "user-1.json",
+    "bodyFileName": "repos_hub4j-test-org_testcreaterelease-2.json",
     "headers": {
       "Server": "GitHub.com",
-      "Date": "Thu, 10 Jun 2021 15:39:39 GMT",
+      "Date": "Thu, 10 Jun 2021 15:53:20 GMT",
       "Content-Type": "application/json; charset=utf-8",
       "Cache-Control": "private, max-age=60, s-maxage=60",
       "Vary": [
         "Accept, Authorization, Cookie, X-GitHub-OTP",
         "Accept-Encoding, Accept, X-Requested-With"
       ],
-      "ETag": "W/\"02702ed486b737dbebf75786b5e6754c51de30f4df9ae2303e75d496711dd609\"",
-      "Last-Modified": "Wed, 09 Jun 2021 22:41:52 GMT",
+      "ETag": "W/\"f6c2944bb9667ef3a8e285429d3578819365ee57eca45b34f24b0260b3b5716a\"",
+      "Last-Modified": "Thu, 10 Jun 2021 01:31:14 GMT",
       "X-OAuth-Scopes": "admin:enterprise, admin:gpg_key, admin:org, admin:org_hook, admin:public_key, admin:repo_hook, delete:packages, delete_repo, gist, notifications, repo, user, workflow, write:discussion, write:packages",
-      "X-Accepted-OAuth-Scopes": "",
+      "X-Accepted-OAuth-Scopes": "repo",
       "X-GitHub-Media-Type": "unknown, github.v3",
       "X-RateLimit-Limit": "5000",
-      "X-RateLimit-Remaining": "4935",
+      "X-RateLimit-Remaining": "4917",
       "X-RateLimit-Reset": "1623340890",
-      "X-RateLimit-Used": "65",
+      "X-RateLimit-Used": "83",
       "X-RateLimit-Resource": "core",
       "Strict-Transport-Security": "max-age=31536000; includeSubdomains; preload",
       "X-Frame-Options": "deny",
@@ -38,10 +38,10 @@
       "X-XSS-Protection": "0",
       "Referrer-Policy": "origin-when-cross-origin, strict-origin-when-cross-origin",
       "Content-Security-Policy": "default-src 'none'",
-      "X-GitHub-Request-Id": "C222:E96D:4193E58:42C2987:60C2323B"
+      "X-GitHub-Request-Id": "C2FC:9EEA:11C275E:1296C75:60C23570"
     }
   },
-  "uuid": "c1b5c491-edb4-4985-90b0-3b027fd3a6a7",
+  "uuid": "7cb5aa21-2b0d-4139-be8c-c12816556e50",
   "persistent": true,
-  "insertionIndex": 1
+  "insertionIndex": 2
 }
